Fix sidechain proposal bug, refactor, gui updates
* Fix sidechain proposal hashID2, was 256 bits and should have been 160.

* Make it so that the RPC and GUI versions of createsidechainproposal
validate hashID1 and hashID2

* Update tests for new sidechain code

* Move uint160S() from tests to src/uint256.h

+++ src/sidechain.h
@@ -43,7 +43,7 @@ struct SidechainProposal {
     std::string sidechainHex;
     std::string sidechainPriv;
     uint256 hashID1;
-    uint256 hashID2;
+    uint160 hashID2;
 
     ADD_SERIALIZE_METHODS
 
@@ -93,7 +93,7 @@ struct Sidechain {
     std::string title;
     std::string description;
     uint256 hashID1;
-    uint256 hashID2;
+    uint160 hashID2;
 
     std::string GetSidechainName() const;
     bool operator==(const Sidechain& a) const;
